added manage user command

+++ Entity/User.php
@@ -381,10 +381,12 @@ SQL;
         if ($this->hasRole($role)) {
             $role = new Role($role);
             $role = Role::findBy("name", $role->getName())->one();
-            EQM::nativeExecute("DELETE FROM roles_users WHERE user_id = :userId AND role_id = :roleId LIMIT 1", [
-                'userId' => $this->id,
-                'roleId' => $role->getId()
-            ]);
+            if ($role instanceof Role) {
+                EQM::nativeExecute("DELETE FROM roles_users WHERE user_id = :userId AND role_id = :roleId LIMIT 1", [
+                    'userId' => $this->id,
+                    'roleId' => $role->getId()
+                ]);
+            }
         }
     }
